I changed both of the arguements in the ellipse
I changed the width in the ellipse from 80 to 180
I changed the height value in the ellipse from 80 to 280
Expected: the ellipse should become higher and wider as
these values have been changed
Actual: As expected the size of the ellipse became much bigger
as the height and width values were changed.
The ellipse looks out of shape now as the values of the
elipse are not the same as the had been i,e 80height
and 80 width.