I added another ellipse to the program
I added another ellipse to the program with the same
dimensions as the previous ellipse

Expected result: Both ellipses should appear within the
window an one or maybe both will move accross the screen
as the mouse moves.

Actual result: Both ellipses appear on the screen but
only one ellipse moves with the cursor as it moves accross
the screen. The other ellipse stays in the top far left of
the window and does not move