I removed the ellipse for a rectangle
I deleted the ellipse from the program and replaced it
with a rectangle of the dimensions (160,30,260,20)

Expected: The rectangle should move accross the screen
just as the ellpise did leaving a trail behind it

Actual result:
The rectangle has been drawn but it does not move. When
the rectangle is clicked on the colour inside it changes
from black to white.This happens as the mouse is click on it
but when the mouse is released the colour within it changes
back to white.