I changed the contents of the program to a rectangle that can change colour when the mouse hovers over it
The program declares four ints x,y,w and h and gives each of
 these a value.These are used as the dimension for the rectangle.
They also are used within the if statement that when the mouse
moves over the rectangle on screen it changes from white to black.
When the mouse move off the rectangle its colour changes back to
white. This is done from within the if statement by saying that if the
mouse is outside the rectangle colour it white,but if (else the mouse
is over it fill the rectangle with the colour back.

Expected result: The rectangle will change colour when the mouses hovers
over it. It will change from white to black and then from black to white
when the mouse exits the rectangle

Actual result: As expected the coulr changes on mouse enter and mouse exit.