Created a new rectangle
I created a new rectangle with the same measurements as the
other rectangle on screen. I gave it different coordinates
using a for the x axis and b for the y axis.

Expected result: A new rectangle will be created in the window
of the same size  as the other rectangle but it has different cordinates
so will appear in a different area in the screen

Actual result:As expected a new rectangle of the same size and shape appears
in the window in a different area. The only unexpected result is that when you
hover over the new rectangle nothing happens but when you hover over the older rectangle both rectangles change colour