I created a new rectangle
I created a new rectangle in the window.Similar to the ellipse before
when you drag the mouse accross the screen the rectangle leaves a trail.

Expected result: A rectangle will appear and as i scroll over the screen
it will leave a trail behind it.

Actual result: As expected it created a new rectangle in the window
and it leaves a trail behind it. The only unexpected result is that
another rectangle appeared in the top left hand side of the window