I added an if else statement
I added an if else statement saying that if the mouse is pressed
the colour within the rectangle changes from white to black. The mouse still
leaves a trail behind it and if you click on the mouse while dragging the
mouse accross the screen it leaves black rectangles mixed into
the white rectangles.

Expected result: The colour of the inside of the rectangle should change
from white to black when the mouse is clicked.

Actual result: The outcome is as expected.