I removed the ellipse and added a rectangle instead
I removed the ellipse and using the same values of the ellipse
I created a rectangle.

Expecteed result: I would hope that the rectangle will appear in the
center of the screen just as the ellipse did.

Actaul result: The rectangle did appear but it did not appear in
the center like I had hoped.The lines that had already been drawn
are now fully visable and and intersect each other in the center of
the window The rectangle now appears just under and to the right of the
center of the window just under where the lines intersect
i